Get ready to include some Twitch code for testing
For Day 2, the desire was to work on Twitch authentication, and to that
end we set up appropriate packages for being able to make requests of
twitch.

The goal was then to work on AUTH, but what we ended up doing was just
playing with some sample code using the passport library, which was
suboptimal.

That code is not included in this repository because it was a sample
written by others, so this day is written off to research.