Force verification every time auth happens
This includes in the initial authorization request a flag which tells
Twitch that it should force a re-authorization request even if the user
has already authorized.

This is important because we may want to swap the user that's
authorized, since this is for a bot and perhaps one might want to run
it as themselves OR a specific bot account and swap between.