Include config powered by convict and dotenv
Include the ability to configure the bot by using convict to have a
configuration schema and dotenv to be able to specify the environment
variables needed to configure things.

The logic here is that the config file we're using is a stub that would
be edited by someone else, but for our purposes during development we
want to be able to change the values without accidentally checking
anything in.

Also, some values such as the client secret are probably best kept not
in the source code, since that is how accidents happen.