Refactor the auth code to a separate file
For clarity in the code, the functionality that relates to the auth
mechanism that we're using has been pushed out to it's own file and
documented to clarify how everything works.

The idea going forward is that a part of this will be to grab out the
auth tokens (if any) and request any we need, and so on.