Create a single object to group our API objects
As development proceeds we're going to keep adding in new modules that
give us the different functionalities that we desire, and that's going
to mean that some call signatures keep getting longer and longer as we
either need to provide them with a subsystem's object OR take one to
pass it through.

This change makes an "api" object that groups all of this as well as
other related information we may need to keep everything logically
together.