Include a simple crypto system
Undoubtedly this could be better, but for the time being implement the
ability to encrypt and decrypt string data. This will be used as a
safeguard to make sensitive data in the database not immediately
obvious to the casual observer.

As a part of this change, the configuration now includes a key for the
encryption system's secret, which is a sensitive value.