Start working on the front end of the bot
Here we're starting to work on the user visible portions of the bot in
order to make it easier to use and control than the traditional method
of knowing to put magic things in places.

This first mock up is for a panel that will show you what account is
being used for the bot (if any) and allow you to log in and out.