Add support for dynamically updating login panel
This is the first part of a two part change, in which we have a new
function that can update the bot authorization panel to indicate that
no account is authorized, and set up the link to allow for that.

More work needs to be done here to also update this based on a given
piece of authorization data from the back end.