Include authorization and twitch API in the API
When the authorization system is initially set up, this will add into
the global API method the authorization object and a Twitch API object
that are related to the account that's authorized for the bot.

In the case where there is no authorized account yet, this will set the
api.auth item to `null` as an indication of this (and api.twitch will
not be defined).