Finalize the login authorization flow
This change sets up the ability of the bot to end-to-end allow you to
authorize and deauthorize a twitch account to be used as the account
the bot runs as.

This involves making sure that the twitch API handlers are created and
removed as needed when authorization and deauthorization happens, as
well as creating a back end deauth routine to drop the token.