Update the back end to support multiple auth
We're going to need the back end code to support the idea of authorizing
more than one account, since we want to not only authorize a bot account
to run the bot as, but also authorize the bot to enter the channel of
the user running the bot.

To that end, the infrastructure on the back end needs to be smart enough
to know about more than one auth type, which is handled currently by
having more than one callback URL, which can be used to disambiguate
them.