Include a panel for the authorized channel
In our current design, we allow the user to authenticate via the account
they want to run the bot as, and then a second authentication is used
to specify what channel the bot should run inside of.

This latter authentication also provides the ability of the bot to take
actions in the desired channel.

This change includes the beginnings of the ability to set the channel
via an authorization.