Fix our issues with authorization state staleness
We had an issue with the existing solution in which the front end would
need to ask the back end for the URL to go to when clicking on the Auth
buttons because they contain information specific to the app that only
the back end knows; the request would also contain a state.

The issue we had was that it was possible in various ways for the back
end and front end to get out of sync with each other regarding the
state, which would make the authorization requests fail.

The simple solution here is to have the front end request a local URL
that collects the correct URL and then redirect the browser there
directly. This allows things to remain in sync.