Use an app token for the Twitch API requestor
Previously we were using the token from the bot account to set up our
Twitch API endpoint for making requests. This is not a great idea for a
few reasons, one of which is that if there's no bot account authorized
then no twitch API can be used, but we always want to use it.

This changes the code such that we now get an App token that we use
for our API requests instead. Any requests that require user specific
permissions will require that the user in question has authorized the
clientID of our app to take those actions.

By strange non-coincidence, that's exactly why we have authorization set
up the way we do.