Set up channel configuration information in the db
This adds in a new table and special handling for tokens created by the
user that authorizes the bot to join a specific channel.

When the user authorizes the bot for their channel, instead of storing
the token (since we don't actually need it) we gather their information
directly and store it in a table so that we can easily get at it.

This needs to be refactored a bit since it's redundant and also storing
data that doesn't need to be stored (such as the description and view
count, which are changeable). Realistically it only needs to store the
user ID and name.