The chat lives!
This includes some hacky, commented out changes that makes the bot
connect to the chat, reading the chat and displaying it to the console
as well as saying something when it first joins.

This needs to be made "not dumb", but a few more minor tweaks are needed
before that's ready.