Factor the Twitch API code out of the Auth code
The authorization code was being used to set up the Auth token because
it used to use the bot token, and that was set up in the Auth code.

Now that this is no longer the case, the Twitch API is now set up
directly within it's own code file.