Swap the channel config table for a user table
The idea of a configuration table that represents the channel was
problematic.

The main reason is that it was storing data about the user that is
dynamic and changeable, which means that most of the fields were
redundant.

Secondly there would only ever be a single channel, so a whole table
devoted to that seems silly.

It has now been removed and in it's place there is a users table that
stores the ID's of user accounts according to their role.

Currently this is just the user whose channel the bot runs in, but that
will be changed to also include the bot account as well.

bugging me, so things have been changed to use a Users table instead.

Here in this table