Perform some light authentication cleanup
This cleans up the auth code slightly and changes the location of the
file as a final step of cleanup. There was some code that is no longer
needed that was redacted away as well.

Further cleanup could be done here to make the flows more correct, but
we can do that as part of a future change.