Include the redacted chat handler
The previous day's last commit was to add in some test chat code to make
sure that things were working as we might like, and the first commit of
today was to remove it.

This brings it back (with some slight changes) so that it's now a proper
part of the bot. At startup, the bot will join the chat, but ONLY if
it has an associated bot account and channel account to work inside of.