Set up events that indicate of auth status change
This change alters the authorization code so that whenever an account is
authorized or deauthorized (either bot or chat), a message is
transmitted to interested listeners to let them know that the change
happened.

This is done when the state actively changes at runtime, and auth events
are also synthesized at startup to allow anything that cars to know when
it's time to proceed.