Prepare for setting up the chat based on events
This builds in the framework we just added and will allow the bot to
join the chat when it knows that both accounts have been authorized and
leave the chat if one of the accounts becomes deauthorized.

This is done by setting up appropriate items in the api to convey this
information and whether or not it's been done. The actual joining of
the chat has been left as an exercise to the user.