Namespace the various chat options
The chat functionality has a few different things that it needs to add
to the API; this moves them all from being separate top level entities
to being sub entities in an overall namespace to make life a little bit
better.

Since some of the names are now a little longer than we would like,
there is now also some aliases used in places to make life better
overall.