Include a command parser in the api struct
This extends the API structure that's passed around the bot so that it
contains a member for parsing commands, so that there is a
single "global" instance that can be used.

This instance is used in the chat code's message handler to parse out
the incoming text in order to determine if it's a command or not.

This is currently semi-stubbed out, and only displayed the parsed data
without taking any other actions.