Include a database schema for the command list
All of the commands in the bot are controlled by a database entry, which
allows for dynamic command creation and for the code in them to be hot
reloaded.

The ultimate goal is to have a bot which can have it's commands amended
without having to touch the source in an editor (in that the front end
will provide such powers for us).