Fix the token schema filename
In a previous change we renamed the schema that was used for
authorization to be used to store tokens instead. The name of the
schema changed, as did the table, but we didn't rename the file. Fixed.

This also fixes some code style issues encountered along the way because
I am apparently terrible and remembering where the guards go and how
they are supposed to be spaced.