Include some new libraries for upcoming changes
In order to complete the next steps, we need to include some extra
libraries that we don't currently have enabled in this project.

These libraries allow us to use filename matching to know when files
need to be hot reloaded, and a library that allows us to make that hot
reload actually happen.