Include our generic handler classes
This brings in a handler class and overall mapping handler that can be
used to associate arbitrary functions with names, and have those
functions be hot re-loadable and executable at will.

This will be the core of our command system as well as our event and
channel point redemption systems, when and if we get to those aspects
in this bot.