Include the concept of roles to the bot
The idea here is that you may often be running a copy of your bot to
work on it while also having a "production" version that is also
running at the same time. In such a case it's frequently annoying that
while executing commands, multiple bots respond when you don't want
them to.

The main way to combat this is to have commands be enabled or disabled
in different versions so that only one one will actually do something,
but occasionally you can't or don't want to do that.

Roles allow a way around this by assigning each bot a specific role
(such as "dev" or "prod") and then being able to execute a command and
direct it specifically to bots in a particular role.