Set up a database table for regular users
One of the permission levels that we want for the bot is that
of "regulars"; people that are frequently in the chat and thus can be
trusted to have more privileges than others.

This sets up a table to track such users. The table stores not only the
userid but also the user name and the display name. Ultimately the id
is what's used to grant the permission; the other data is subject to
being wrong because users can change their account names.

Some post process could be done to periodically update user names if
so desired.