Include a core command to handle the regulars list
One of the core commands in the bot is "regulars", which allows you to
modify the list of users that are frequent enough in the chat that you
wish to allow them special permissions, such as running commands or
being exempt from some things.

This adds in the command and also fixes the `api.getUserlevel
()` function to access it.