Include the ability to bootstrap core data items
The bot allows for dynamic items to be added to lists (such as commands)
but there are also a set of core items that should always be present.

This change includes a bootstrap step that ensures that all such items
are present in the database at startup by including them via a template
item.

Any items that are found to exist are left alone, so that any changes
made to their properties are not clobbered.