Modify the handler map to know about reloads
The handler map's `reload()` method is for dynamically reloading source
files to bring new command changes and the like into existence.

This change modifies it so that it returns a value that indicates
whether or not it actually found something to reload or not, or if an
error occurred.

This can be used by callers to tell if something actually happened or
not.