Include a core command to do reloads
This brings in a core command that can leverage the handler map's
ability to hot reload files to allow for commands to be hot reloaded
at will.

This is probably best kept in its own file so that an error while doing
a hot reload doesn't have the ability to knock out the reload
functionality unless you're trying to modify the reload command and get
something wrong.