Include a new command for setting the bot role
This brings in a new command that allows you to query and change the
value of the default role that is assumed in all commands that don't
have a role= argument that specifies it.

This allows you to indicate to the bot that unless you tell it
otherwise, all commands are directed to a specific role only.