Augment how the api.chat.say helper works
This changes the api.chat.say helper alias so that it's set up by a
function instead of being inlined. This is because it will now log to
the console after it says something so that the logs of the chat also
include what the bot said.

This makes debugging easier when going through logs because we can see
how the bot responded.