Clean up the display on the debug command
This changes how the debug command displays things so that it's a little
more clear what all of the sections are.

Additionally, the command now sets up a flag that the command dispatcher
uses so that it will log all messages and not just the output of a
particular command.

This can be enhanced further by making it controllable what portions of
the messages are logged.