Add the ability to perform actions as the bot
This adds a new chat helper similar to `api.chat.say` which does an
action instead of just sending out text, and is named `api.chat.do`.

In addition, a new internal nodecg message has been added to allow for
other systems in the bot to send out a message as an action using the
new helper, as well as augmenting the message that does a text send to
also support replying.