Include some initial work on an edit command
This command would allow you to be able to edit the access level, state
and cooldown of a given command, as well as see the values of those for
any given command.

This is still heavily a WIP as we ran out of time on the stream and in
the day.