When loading a dynamic file fails, show the stack
This is a slight augmentation to the module hot loader for dynamic list
items that displays a stack trace if there's an error encountered while
loading and parsing the source file.

Previously this would only show the error message, but frequently this
is not enough information to be able to locate the error.