Include some helpers for upcoming changes.
This includes three helper functions to be used by most/all of the
commands in the editing code.

The first is for getting a target command and verifying that it follows
specific rules, such as existing, being or not being an alias or whether
it is a core command or not.

The second is for storing changes made to a command record, which will
update the database and also patch the currently loaded command record.

The third is a visual helper for being able to show cooldown times in a
human readable way. This last one is not currently very good, but it's
good enough for our purposes here.