Include the enable and disable commands
This implements the command and the alias on the command that allow the
user to control whether or not commands are enabled without having to
modify the database directly.

This currently refuses to modify the state of a core command to avoid
issues like disabling the command that enables commands, locking you out
of commands.