Include the access level change command
This includes a command that lets you easily check the access level for
a command as well as dynamically changing it.

This ensures that the access level specifies is correct, and while it
has a list of valid levels that it shows the user, the list it actually
draws on is much larger and allows for more freedom.