Tweak the default user levels for core commands
This changes the default access levels for some of the core commands so
that the ones that logically should also be accessible to the mods in
the channel as well.

These are "safe" in that the core commands themselves won't let anyone
modify any of their information, meaning that a mod can't anything too
unseemly in the core (though obviously, trust your mods).