Include the start of a command to modify aliases
This includes the beginnings of a command that can view, add and remove
aliases for commands.

This version is still a WIP, but adding aliases is now working (though
perhaps it is not updating the local object properly and duplicating
things, it's a little unclear at the moment).

This needs to be cleaned up and the alias removal needs to be included
as well.