Finally fix the display of the chat controls
THanks to the generous support from someone in chat who provided a link
to a jsfiddle that had a minimal example of flex, the controls for the
chat are now properly sized and aligned within their panel.

This may or may not be properly responsive, but that is not a concern
at the present time since the UI needs further work in the future.