Rename the file the file the command edits are in
This renames the core file that the commands for editing other commands
are stored inside of; the other name didn't fully make sense for what
the contents are.

Currently the command for viewing command information is still in the
same file when it should probably be moved, but that would require some
code refactoring so we're not going to worry about that just yet.