Update the handler list to support dynamic aliases
Although the handler map in general handles aliases dynamically on load
and reload of files, making sure that aliases are set up and torn down
automatically, there is a need (for example in bot commands) to adjust
the list dynamically.

This change adds in an API that allows for this to happen. Changes that
are made to the list at runtime change only the list; so the underlying
item doesn't get told that it has more or fewer aliases; that is the
job of whatever is invoking the method.

Additionally, although the method generates warning messages and will
make sure not to break the list, it doesn't display anything on success
under the theory that whoever is changing the aliases may have a custom
desire to do so.