Finalize the initial command to add aliases
This finishes up our initial take on the command that's used to add
aliases at runtime. The one piece that was missing from our previous
solution was that although the database and the command itself got
updated, the list of commands did not.

This now uses the new API in the handler list to take this action.