Finalize the initial command to remove aliases
This finishes the work from prior changes to allow for the alias command
to remove aliases as well.

This also includes some fixes for the alias addition, which was working
but in a way that the removal would not because the addition relied on
the first alias in the changes list being the new one, which is not
something that the remove can do.

This allows us to move the newly added alias to the end of the list,
where one might actually expect it to live.