Set the option to assume that the bot is a bot
The Twurple chat client has rate limiters in place to throttle traffic
to Twitch. Mods get a different rate limit than normal users to.

When initializing the chat client, we can indicate that the bot is
always a mod, which changes its rate limiter.

This option is safe to include, but ONLY if you make your bot a mod;
so we should probably point out that this is required.