Allow the module loader to propagate load errors
Previously if there was a load error of any kind, either at startup or
during a manual runtime hot reload, any errors would be caught and
displayed in the console.

This change allows the system to propagate those back out to the caller
so that they can tell if there were any problems, and do more
interesting things with them, such as display them in the dashboard.