Fix the error where changing a filename loses cmds
The reload mechanism used to reload files based on first unloading all
commands that match the list of files being reloaded, and then finding
all items in the changed files to add them back.

This causes a failure if the name of the file that a command is stored
in changes, since it will now no longer be reloaded unless it now
happens to be in a file that is also being reloaded for other reasons.

This has been changed so that instead we look up information for each
unloaded command and then load based on that, which will give us the
new command (if it exists still) even if the file name changes.