Add a new commands workspace and log panel
This change adds to the list of available dashboard panels a logging
panel for the command loader which can display the results of both the
initial load of commands as well as the results of any reload.

This is meant to display the same information as we see in the console,
but in the web front end so that if the bot reports an error, we can
see it without having to go to the console.

This lives in a new workspace named Commands, which will be the eventual
home of items that allow us to see the existing commands at a glance.