Include a new command for adding new commands
As meta as this sounds, this command allows us to automatically create a
new command by adding a new entry to the database and then, if the
source file in question does not exist, putting a new stub in place.

Once the file is in place (either new or not), it's reloaded to bring
the new command into the system, although in the case of a newly added
file from a stub, some editing will be required.