Include  new command for removing existing cmds
This is the analogue of the previous addition and allows us to remove a
command from the system. When removed the database is the only thing
that is modified; the source file remains behind.

The command stops you from removing a core command, since that can get
you into a bit of a sticky widget.