Include a new utils lib with a cmd usage helper
This includes a new utils file that is meant to hold commonly used code
that is used in multiple places.

The first entry is a function that will make it easier to get consistent
usage information for commands by abstracting it away into a single
function that takes arguments that will then be arranged into an
appropriate message and sent out to the bot.