Slightly change up the command schema to use ints
The booleans in the command schema were changed so that their default
values are now either 1 or 0 instead of true and false.

Behind the scenes Trilogy stores 1 and 0 when you make changes, but the
defaults are used to populate new entries and that makes the data not
look consistent, particularly if you're not using Trilogy to fetch it.