Create a new schema for automatic text responses
Create a new table for storing simple text responses that should come
from commands when they're executed. This is meant to be linked to a
command that would look up an entry and display it, such as the ever
popular "!project" command that would display the current project.

Such commands are useful and ubiquitous but would be a pain to set up
using code. So instead, we have a table that associates messages with
alias names, and an upcoming command will use the data from the table
to respond as appropriate.