Include a simple stub version of the text command
This command is used to both set up and manipulate the list of auto
entries in the system, as well as to use them to get the bot to say
predefined segments of text.

The version here implements the user facing side that triggers the
messages, but doesn't include the broadcast facing side that lets you
set them up.