Alter the name and structure of the responders
As part of an upcoming change to text responses, this alters the name
and layout of the table used to represent automatic text responders so
that they can appear to operate more like standard commands.

In use a special dispatcher in the bot will trigger them after looking
them up, similar to how commands work.

Despite the structures that are used in this operation, commands and
text responders have different namespaces, so it is possible for them
to share names (although in this case the responder will be unavailable
as commands take priority).