Include a new static handler list item
This includes a new class that's a subclass of the standard handler
class, but which is meant to wrap static items; that is, items for
which instead of a source file that's loaded dynamically instead the
items in the list have a native handler.

This is used for things like the responder items, which just need to
say text and as such don't need customized code.