Update the text command stubs
Previously the text command was set up to both trigger a response and
also edit it, with the editing being very roughly stubbed out.

With the newly changed code, we instead have a different method that is
now used to generate the responses.

So, this change alters things such that the operations for the text
command are now stubbed out more completely, and the code for generating
responses is removed.