Set up a back end web API to serve and save files
The next task is a file editor from directly within the dashboard, and
to get that working we first need an API that allows for fetching and
storing the contents of files, as well as being able to query command
details.

This sets up just such an API, currently only for commands, which allows
the front end to see the list of commands as well as being able to not
only fetch the files that implement commands, but also put them back
after they're edited.