Include a simple text editor library and font
For the sake of simplicity and for (seemingly) better mobile device
compatibility, include the Ace text editing component for use in the
online editor. This powers a few different online services and is fairly
capable as an editor as these things go.

This also includes the same Iosevka font that we normally use, converted
into a WebFont version for use in the editor online.