Include the initial editor component
Include a brand new editor panel that uses the Ace editor that we just
added to the project and the web API that was recently added to allow
the user to look at and edit the contents of any of the command files
in the bot from directly inside of the dashboard.

This will fetch the list of commands, sort them out according to the
file that they're stored in, and set up some select boxes such that
when the current item changes, we load an appropriate file.

This also includes a save function that can send the file back, and a
checkbox which can trigger a reload of a named item so that the back
end will reload the contents of the file.