Modify the discard button to also do a reload
Previously the discard button was only enabled when there were changes
in the buffer, and would revert the contents of the buffer by first
prompting, and then re-requesting the file. When the file was not
dirty, the button would disable itself.

This has been changed so that the button is always enabled and instead
changes its caption to be either Reload or Discard, so that if you
leave the panel open and come back later, you can ensure that you have
the latest version of the file that there is.