Factor the Twitch fullbleed out into panels
Previously there was a single fullbleed Twitch panel that contained both
of the authorization displays as well as a small section for sending
messages.

This has been exploded out into two different workspaces with everything
in their own panels instead, for clarity.

The Auth workspace now has individual panels for the two items which
could contain authorization information and a new Bot panel contains
the controls for sending messages.

This requires a small change to the back end so that the authorization
code will redirect to the correct place, since the workspace has changed
and is no longer a full bleed.