Upgrade scopes and handlers to support whispers
This alters the permission scopes that we ask for so that we can
(in theory) send and receive whispers. A handler is also set up for the
purposes of logging when a whisper arrives.

As of the current time, although we can receive whispers, the bot seems
unable to send them (as either the test account or the more long lived
bot account) although the account in general can, if logged in via a
browser.