Include the google tts library
This library abstracts away all of the work needed to get Google's TTS
system up and running; it requires that you have an account that has the
TTS options turned on.

TTS is something that technically costs money after a certain number of
characters of text have been sent, BUT the limits are rather high for
the purposes of a streamer, and they're even higher if you don't use one
of the higher quality WaveNet synths.