Include configuration options for working with TTS
To start the process of bringing in Text To Speech using Google's TTS
API, this includes the configuration records that are needed in order to
set up the voice, gender and language that the bot uses.

There is also a fourth configuration option which is used by the Google
libraries that tells it where to get the appropriate file with the
information it needs for authorization, but for now that's handled by
that library, so you just need to set up an extra environment variable.