Include the TTS back end core
Set up the core back end required to do text to speech. Our system uses
a job system wherein asking the system to speak text queues up a job,
gives it a unique ID, and then asks the front end code to play it.

The front end will request that the back end perform the job, which will
cause it to talk to google and proxy the audio back so that the overlay
can use it as the src of an audio tag.