Add in the graphic overlay for the TTS system
This sets up the page that runs inside of OBS or your streaming software
of choice and listens for messages from the back end saying that it's
time to say text.

When one is received, we ask the back end to give us the audio, which it
proxies from Google's TTS API for us to play back.