Include a simple first pass UI on the TTS system
To bring everything together, this includes a new panel in the dashboard
for talking to the TTS system and getting it to speak text.

This is currently done with a series of dropdowns, though in reality it
should be fewer and have filters that update following items so that
the list of appropriate choices is always correct.