Include a voice data file
Include a JSON file that contains all of the possible English language
voices that the Google TTS system supports. This contains all of the
fields for that language, plus an addition one for use in the front end.

This was gathered by using the TTS system's endpoint for gathering
voices based on a particular language, which we could also do here.
However for the sake of simplicity in the example code, we just hard
code the data into a file instead.