Add a new endpoint on the back end tts for voices
This includes a new route that allows for fetching the list of voices
that was previously added in a JSON file. The request will load the file,[
parse it using the JSON5 library (so that it's more complacent about
possible errors) and then return it back.

This also changes the endpoint for actually speaking text, which has
been also adjusted in the front end, which is what calls it to make the
speech magic happen.