Adjust the front end to use the new voices API
This includes a simple change wherein we ask at load time for the list
of voices, and then auto-populate the voice list using them.

This filters the voices in the select tag so that only those voices that
match the language and gender selected from the other tags are allowed,
since Google will happily let you mix and match and then not do what
you expect.