Factor some code out to be common utility code
There is some code for doing conversions on cooldown values and user
levels which were contained wholly in one of the commands files, but
which will be of use in at least one place, if not others.

These were factored out into the utils file so that they can be easily
reused, and were also renamed to make it clearer what they're for,
since they now live in a more project-wide context.