Move our runtime data files to a folder
This moves files (currently only one) to a new `data` folder that is
meant to hold all of the runtime data files that may be needed.

The configuration file is not included in this despite the fact that
it's needed at runtime, because it seems weird to consider your config
file a data file.