Refactor the chat code slightly
Previously the code for handling incoming chat messages was inlined in
the event handler's body for incoming messages. While this works, it is
a bit of a pain for finding and modifying, and since it's the most
complex part of handling a message, also makes things harder to read.

This change factors the code for handling things out into a set of
functions instead, which allows for a clearer delineration of what's
happening and why.