Fix the name of the responders schema file
During the period where we were waffling on what to call the name of the
schema for auto responders, we named the file responses and then
decided that the name responders was better, but didn't fix the file.

Since that is very nerve wracking and breaks the pattern, that is now
fixed.