Include a new utility function to dedent text
There are a few places where we use template strings to capture longer
strings of data because they can span multiple lines more easily than
"standard" strings.

This causes a bit of an issue in places where strings are used for
logging or other places where we want them to convey as a single string
despite being split in the code for readability.

This includes a simple function that does this for us to make our lives
easier and less repetitive.