Factor the command prefix list into the utils code
Previously the list of characters that can be a command or responder or
alias prefix was held in the parsing code as a class static.

This needs to be used in a variety of places, such as when verifying
that a command name is valid or just to specify what could be added.

For clarity purposes, this has been factored out into the utility code.

Since there will soon be more than one place where an inferred default
command prefix will be referenced, this also includes a new config item
that allows you to specify what the default prefix is.