Streamline the bootstrapper
This is a slight improvement on the bootstrapping of the database at
startup, which it would appear on windows is hideously slow though it is
near instantaneous on Linux.

Instead of executing all of the queries serially, we instead kick them
all off async and then wait for them all to finish, which makes it about
4x faster for a complete re-bootstrap.

A better approach would be to use a single query to grab all items and
then use simultaneous async requests to insert only those which are
needed, but that's being deferred here for the time being due to time
constraints.