Refactor some of the command code
There are a couple of bits in the command code which are likely going to
be useful in our upcoming text responder command.

These have been factored to both have a bit more of a pleasing name and
also to move them into the common utils code so that they can be seen
in multiple places.