Assume seconds for cooldowns with no suffix
Previously the code enforced that the suffix on a cooldown timeout was
one of 's', 'm' or 'h' and would generate an error if not.

This is now changed so that if the suffix isn't one of these, then it's
assumed to be `s` so that bare numbers are allowed.