Make command name helpers handle missing commands
The helper items for both verifying that a command name is valid and for
making one valid if it's already valid both previously assumed that
they would be given a string which represents a command name, and act
accordingly.

Although this is a good general assumption, there are some cases were
optional arguments may be provided that are names, and in such a case
the code gets more complicated if we need to first check to see if it's
there before calling one of these.

So, both are now adjusted such that a command name that's undefined is
not valid, and trying to make it valid just directly re-returns
undefined without doing anything.