Make it easier to view and change auth scopes
The authorization mechanism needs to provide a list of scopes that
represent the specific permissions to data and actions that are being
requested, so that the user knows the full scope of what the code can
do on their behalf.

This change makes those scopes easier to see and change for both the
bot and user accounts by making them easy to see and edit in the file
as a top level constant instead of being embedded directly in the main
code for authorization.