Update the auth system to provide userID in events
This is a simple change that allows the events for authorization and
deauthorization to transmit along with the type of event the userID of
the affected user.

This is something that is not needed for chat, but will be needed for
upcoming event changes, wherein we need to know the user for whom we are
removing or adding events.